Deduplicate filename-to-language detection into common crate
Remove the duplicate get_language_for_file, get_extensions_for_language,
get_exact_filename_for_language, get_prefix_for_language, and their
backing FILE_*_PER_LANGUAGE_LIST tables from cli::file_utils, keeping a
single copy in common::model::language. cli now delegates to common's
implementation, and the associated tests are migrated there too.
